refactor(renderer): establish app boundaries
