+++ src/renderer/src/apps/chat-main/ChatMainApp.vue
@@ -1,0 +1,616 @@
+<script setup lang="ts">
+import { onMounted, ref, watch, onBeforeUnmount, computed } from 'vue'
+import { useEventListener } from '@vueuse/core'
+import { RouterView, useRoute, useRouter } from 'vue-router'
+import { createConfigClient } from '@api/ConfigClient'
+import { createOnboardingClient } from '@api/OnboardingClient'
+import SelectedTextContextMenu from '@/components/message/SelectedTextContextMenu.vue'
+import { useArtifactStore } from '@/stores/artifact'
+import { useSessionStore } from '@/stores/ui/session'
+import { useAgentStore } from '@/stores/ui/agent'
+import { useDraftStore, type StartDeeplinkPayload } from '@/stores/ui/draft'
+import { usePageRouterStore } from '@/stores/ui/pageRouter'
+import { Toaster } from '@shadcn/components/ui/sonner'
+import { useToast } from '@/components/use-toast'
+import { useUiSettingsStore } from '@/stores/uiSettingsStore'
+import { useThemeStore, type ThemeMode } from '@/stores/theme'
+import { useLanguageStore } from '@/stores/language'
+import { useI18n } from 'vue-i18n'
+import TranslatePopup from '@/components/popup/TranslatePopup.vue'
+import ModelCheckDialog from '@/components/settings/ModelCheckDialog.vue'
+import { useModelCheckStore } from '@/stores/modelCheck'
+import MessageDialog from '@/components/ui/MessageDialog.vue'
+import McpSamplingDialog from '@/components/mcp/McpSamplingDialog.vue'
+import { initAppStores, useMcpInstallDeeplinkHandler } from '@/lib/storeInitializer'
+import { ensureIconsLoaded } from '@/lib/iconLoader'
+import 'vue-sonner/style.css' // vue-sonner v2 requires this import
+import { useFontManager } from '@/composables/useFontManager'
+import AppBar from '@/components/AppBar.vue'
+import { useDeviceVersion } from '@/composables/useDeviceVersion'
+import WindowSideBar from '@/components/WindowSideBar.vue'
+import SpotlightOverlay from '@/components/spotlight/SpotlightOverlay.vue'
+import { useSpotlightStore } from '@/stores/ui/spotlight'
+import { useSidepanelStore } from '@/stores/ui/sidepanel'
+import { useSidebarStore } from '@/stores/ui/sidebar'
+import { useAppIpcRuntime } from '@/composables/useAppIpcRuntime'
+import {
+  clearGuidedOnboardingResumeIntent,
+  GUIDED_ONBOARDING_RESUME_REQUESTED_EVENT,
+  readGuidedOnboardingResumeIntent,
+  type GuidedOnboardingResumeRequestDetail,
+  type GuidedOnboardingResumeTrigger
+} from '@/lib/onboardingResume'
+import type { GuidedOnboardingStepId } from '@shared/contracts/routes'
+import type { DatabaseRepairSuggestedPayload } from '@shared/types/databaseSchema'
+import { createWindowClient } from '@api/WindowClient'
+
+const DEV_WELCOME_OVERRIDE_KEY = '__deepchat_dev_force_welcome'
+
+const route = useRoute()
+const configClient = createConfigClient()
+const onboardingClient = createOnboardingClient()
+const windowClient = createWindowClient()
+const artifactStore = useArtifactStore()
+const sessionStore = useSessionStore()
+const agentStore = useAgentStore()
+const draftStore = useDraftStore()
+const pageRouterStore = usePageRouterStore()
+const sidepanelStore = useSidepanelStore()
+const sidebarStore = useSidebarStore()
+const spotlightStore = useSpotlightStore()
+const { toast } = useToast()
+const uiSettingsStore = useUiSettingsStore()
+const { setupFontListener } = useFontManager()
+setupFontListener()
+
+const { isWinMacOS, isMacOS } = useDeviceVersion()
+
+watch(
+  isMacOS,
+  (mac) => {
+    if (typeof document === 'undefined') return
+    document.documentElement.dataset.platform = mac ? 'darwin' : 'other'
+  },
+  { immediate: true }
+)
+
+const themeStore = useThemeStore()
+const langStore = useLanguageStore()
+const modelCheckStore = useModelCheckStore()
+const { t } = useI18n()
+const toasterTheme = computed(() =>
+  themeStore.themeMode === 'system' ? (themeStore.isDark ? 'dark' : 'light') : themeStore.themeMode
+)
+// Error notification queue and currently displayed error
+const errorQueue = ref<Array<{ id: string; title: string; message: string; type: string }>>([])
+const currentErrorId = ref<string | null>(null)
+let errorDisplayTimer: number | null = null
+
+const { setup: setupMcpDeeplink, cleanup: cleanupMcpDeeplink } = useMcpInstallDeeplinkHandler()
+
+const resolveThemeName = (themeMode: ThemeMode, isDark: boolean) => {
+  return themeMode === 'system' ? (isDark ? 'dark' : 'light') : themeMode
+}
+
+const syncAppearanceClasses = (themeName: string, fontSizeClass: string) => {
+  if (typeof document === 'undefined') {
+    return
+  }
+
+  const root = document.documentElement
+  // 切换期间临时禁用过渡，让主题瞬时生效，避免大量元素同时跑颜色过渡造成的重绘卡顿
+  root.classList.add('dc-theme-switching')
+
+  for (const target of [root, document.body]) {
+    target.classList.remove('light', 'dark', 'system')
+    target.classList.add(themeName)
+    target.classList.remove('text-xs', 'text-sm', 'text-base', 'text-lg', 'text-xl', 'text-2xl')
+    target.classList.add(fontSizeClass)
+  }
+
+  // 强制同步重算，使本次类切换在「过渡已禁用」的状态下完成
+  void root.offsetWidth
+  // 下一帧恢复过渡，不影响日常 hover 等交互动画
+  requestAnimationFrame(() => {
+    root.classList.remove('dc-theme-switching')
+  })
+}
+
+watch(
+  [() => themeStore.themeMode, () => themeStore.isDark, () => uiSettingsStore.fontSizeClass],
+  ([themeMode, isDark, fontSizeClass]) => {
+    const nextThemeName = resolveThemeName(themeMode, isDark)
+    syncAppearanceClasses(nextThemeName, fontSizeClass)
+  },
+  { immediate: true }
+)
+
+// Handle error notifications
+const showErrorToast = (error: { id: string; title: string; message: string; type: string }) => {
+  // Check if error with same ID already exists in queue to prevent duplicates
+  const existingErrorIndex = errorQueue.value.findIndex((e) => e.id === error.id)
+
+  if (existingErrorIndex === -1) {
+    // If there's currently an error being displayed, add new error to queue
+    if (currentErrorId.value) {
+      if (errorQueue.value.length > 5) {
+        errorQueue.value.shift()
+      }
+      errorQueue.value.push(error)
+    } else {
+      // Otherwise display this error directly
+      displayError(error)
+    }
+  }
+}
+
+// Display specified error
+const displayError = (error: { id: string; title: string; message: string; type: string }) => {
+  // Update currently displayed error ID
+  currentErrorId.value = error.id
+
+  // Show error notification
+  const { dismiss } = toast({
+    title: error.title,
+    description: error.message,
+    variant: 'destructive',
+    onOpenChange: (open) => {
+      if (!open) {
+        // Also show next error when user manually closes
+        handleErrorClosed()
+      }
+    }
+  })
+
+  // Set timer to automatically close current error after 3 seconds
+  if (errorDisplayTimer) {
+    clearTimeout(errorDisplayTimer)
+  }
+
+  errorDisplayTimer = window.setTimeout(() => {
+    // Handle logic after error is closed
+    dismiss()
+    handleErrorClosed()
+  }, 3000)
+}
+
+// Handle logic after error is closed
+const handleErrorClosed = () => {
+  // Clear current error ID
+  currentErrorId.value = null
+
+  // Display next error in queue (if any)
+  if (errorQueue.value.length > 0) {
+    const nextError = errorQueue.value.shift()
+    if (nextError) {
+      displayError(nextError)
+    }
+  } else {
+    // Queue is empty, clear timer
+    if (errorDisplayTimer) {
+      clearTimeout(errorDisplayTimer)
+      errorDisplayTimer = null
+    }
+  }
+}
+
+const router = useRouter()
+const activeTab = ref('chat')
+const isStartupRouteReady = ref(false)
+const processingStartDeeplinkToken = ref<number | null>(null)
+const processedStartDeeplinkToken = ref<number | null>(null)
+
+const isDevWelcomeOverrideEnabled = () => {
+  if (!import.meta.env.DEV) return false
+
+  try {
+    return window.sessionStorage.getItem(DEV_WELCOME_OVERRIDE_KEY) === '1'
+  } catch {
+    return false
+  }
+}
+
+const ensureStartupWelcomeState = async () => {
+  try {
+    await router.isReady()
+
+    const currentRoute = router.currentRoute.value
+    const isWelcomeRoute = currentRoute.name === 'welcome' || currentRoute.path === '/welcome'
+
+    if (isDevWelcomeOverrideEnabled()) {
+      if (!isWelcomeRoute) {
+        await router.replace({ name: 'welcome' })
+      }
+      return
+    }
+
+    const initComplete = Boolean(await configClient.getSetting('init_complete'))
+    let onboardingState: Awaited<ReturnType<typeof onboardingClient.getState>> | null = null
+
+    try {
+      onboardingState = await onboardingClient.getState()
+    } catch (error) {
+      console.warn('[App] Failed to load onboarding state during startup:', error)
+    }
+
+    if (onboardingState?.status === 'completed') {
+      if (isWelcomeRoute) {
+        await router.replace({ name: 'chat' })
+      }
+      return
+    }
+
+    if (!initComplete || onboardingState?.status === 'active') {
+      if (!initComplete && onboardingState?.status !== 'active') {
+        try {
+          onboardingState = await onboardingClient.start()
+        } catch (error) {
+          console.warn('[App] Failed to start onboarding during startup:', error)
+        }
+      }
+
+      if (!isWelcomeRoute) {
+        await router.replace({ name: 'welcome' })
+      }
+      return
+    }
+
+    if (isWelcomeRoute) {
+      await router.replace({ name: 'chat' })
+    }
+  } finally {
+    isStartupRouteReady.value = true
+  }
+}
+
+// Handle font scaling
+const handleZoomIn = () => {
+  // Font size increase logic
+  const currentLevel = uiSettingsStore.fontSizeLevel
+  uiSettingsStore.updateFontSizeLevel(currentLevel + 1)
+}
+
+const handleZoomOut = () => {
+  // Font size decrease logic
+  const currentLevel = uiSettingsStore.fontSizeLevel
+  uiSettingsStore.updateFontSizeLevel(currentLevel - 1)
+}
+
+const handleZoomResume = () => {
+  // Reset font size
+  uiSettingsStore.updateFontSizeLevel(1) // 1 corresponds to 'text-base', default font size
+}
+
+// Handle creating new conversation
+const handleCreateNewConversation = async () => {
+  try {
+    await sessionStore.startNewConversation({ refresh: true })
+  } catch (error) {
+    console.error('Failed to create new conversation:', error)
+  }
+}
+
+// Removed GO_SETTINGS handler; now handled in main via tab logic
+
+const activatePendingStartDeeplink = async () => {
+  const pendingStartDeeplink = draftStore.pendingStartDeeplink
+  if (!pendingStartDeeplink || !isStartupRouteReady.value) {
+    return
+  }
+
+  const token = pendingStartDeeplink.token
+  if (processingStartDeeplinkToken.value === token || processedStartDeeplinkToken.value === token) {
+    return
+  }
+
+  processingStartDeeplinkToken.value = token
+
+  try {
+    const initComplete = Boolean(await configClient.getSetting('init_complete'))
+    if (!initComplete) {
+      return
+    }
+
+    await router.isReady()
+    if (router.currentRoute.value.name !== 'chat') {
+      await router.push({ name: 'chat' })
+    }
+
+    agentStore.setSelectedAgent('deepchat')
+    if (sessionStore.hasActiveSession) {
+      await sessionStore.closeSession()
+      processedStartDeeplinkToken.value = token
+      return
+    }
+
+    pageRouterStore.goToNewThread({ refresh: true })
+    processedStartDeeplinkToken.value = token
+  } finally {
+    if (processingStartDeeplinkToken.value === token) {
+      processingStartDeeplinkToken.value = null
+    }
+  }
+}
+
+const handleStartDeeplink = (_event: unknown, payload?: Omit<StartDeeplinkPayload, 'token'>) => {
+  if (!payload?.msg) {
+    return
+  }
+
+  draftStore.setPendingStartDeeplink({
+    msg: payload.msg,
+    modelId: payload.modelId ?? null,
+    systemPrompt: payload.systemPrompt ?? '',
+    mentions: Array.isArray(payload.mentions) ? payload.mentions : [],
+    autoSend: Boolean(payload.autoSend)
+  })
+  void activatePendingStartDeeplink()
+}
+
+const handleDatabaseRepairSuggested = (payload: unknown) => {
+  const repairPayload = payload as DatabaseRepairSuggestedPayload | undefined
+  if (!repairPayload) {
+    return
+  }
+
+  toast({
+    title: t(repairPayload.title),
+    description: t(repairPayload.message, {
+      reason: t(`settings.data.databaseRepair.reasons.${repairPayload.reason}`)
+    }),
+    action: {
+      label: t('settings.data.databaseRepair.toastAction'),
+      onClick: () => {
+        void configClient.openSettings({
+          routeName: 'settings-database',
+          section: 'database-repair'
+        })
+      }
+    }
+  })
+}
+
+const handleStartGuidedOnboardingDev = async () => {
+  if (!import.meta.env.DEV) {
+    return
+  }
+
+  try {
+    clearGuidedOnboardingResumeIntent()
+    await onboardingClient.start({
+      force: true,
+      stepId: 'select-provider'
+    })
+
+    if (router.currentRoute.value.name !== 'welcome') {
+      await router.replace({ name: 'welcome' })
+    }
+  } catch (error) {
+    console.warn('[App] Failed to start guided onboarding from dev trigger:', error)
+  }
+}
+
+const CHAT_GUIDED_ONBOARDING_STEP_IDS = new Set<GuidedOnboardingStepId>([
+  'switch-agent',
+  'switch-model',
+  'first-chat'
+])
+
+const routeToGuidedOnboardingStep = async (stepId: GuidedOnboardingStepId | null) => {
+  if (stepId && CHAT_GUIDED_ONBOARDING_STEP_IDS.has(stepId)) {
+    if (router.currentRoute.value.name !== 'chat') {
+      await router.replace({ name: 'chat' })
+    }
+
+    pageRouterStore.goToNewThread({ refresh: true })
+    return
+  }
+
+  if (router.currentRoute.value.name !== 'welcome') {
+    await router.replace({ name: 'welcome' })
+  }
+}
+
+const handleResumeGuidedOnboarding = async (trigger: GuidedOnboardingResumeTrigger) => {
+  const resumeIntent = readGuidedOnboardingResumeIntent()
+  if (!resumeIntent || resumeIntent.trigger !== trigger) {
+    return
+  }
+
+  try {
+    const onboardingState = await onboardingClient.getState()
+
+    if (onboardingState.status !== 'active') {
+      clearGuidedOnboardingResumeIntent()
+      if (onboardingState.status === 'completed') {
+        if (router.currentRoute.value.name !== 'chat') {
+          await router.replace({ name: 'chat' })
+        }
+
+        pageRouterStore.goToNewThread({ refresh: true })
+      }
+      return
+    }
+
+    clearGuidedOnboardingResumeIntent()
+    await routeToGuidedOnboardingStep(onboardingState.currentStepId)
+  } catch (error) {
+    console.warn('[App] Failed to resume guided onboarding:', error)
+  }
+}
+
+const handleGuidedOnboardingResumeRequested = (event: Event) => {
+  const detail = (event as CustomEvent<GuidedOnboardingResumeRequestDetail>).detail
+  if (!detail?.trigger) {
+    return
+  }
+
+  void handleResumeGuidedOnboarding(detail.trigger)
+}
+
+const { setup: setupAppIpcRuntime, cleanup: cleanupAppIpcRuntime } = useAppIpcRuntime({
+  handleStartDeeplink: (payload) => {
+    handleStartDeeplink(undefined, payload as Omit<StartDeeplinkPayload, 'token'> | undefined)
+  },
+  handleStartGuidedOnboardingDev,
+  handleWindowFocused: () => handleResumeGuidedOnboarding('window-focus'),
+  showErrorToast,
+  handleDatabaseRepairSuggested,
+  handleZoomIn,
+  handleZoomOut,
+  handleZoomResume,
+  handleCreateNewConversation,
+  handleToggleSidebar: () => {
+    sidebarStore.toggleSidebar()
+  },
+  handleToggleWorkspace: () => {
+    if (pageRouterStore.currentRoute !== 'chat' || !pageRouterStore.chatSessionId) {
+      return
+    }
+
+    sidepanelStore.toggleWorkspace(pageRouterStore.chatSessionId)
+  },
+  openSpotlight: () => {
+    spotlightStore.openSpotlight()
+  },
+  handleSystemNotificationClick: (msg) => {
+    let sessionId: string | null = null
+
+    if (typeof msg === 'string' && msg.startsWith('chat/')) {
+      const parts = msg.split('/')
+      if (parts.length === 3) {
+        sessionId = parts[1]
+      }
+    } else if (msg && typeof msg === 'object' && 'threadId' in msg) {
+      sessionId = (msg as { threadId?: string }).threadId ?? null
+    }
+
+    if (sessionId) {
+      void sessionStore.selectSession(sessionId)
+    }
+  },
+  getCurrentRouteName: () => router.currentRoute.value.name
+})
+
+// Handle ESC key - close floating chat window
+const handleEscKey = (event: KeyboardEvent) => {
+  if (event.key === 'Escape') {
+    void windowClient.closeFloatingCurrent()
+  }
+}
+
+// Same listeners as before; VueUse owns register/unregister with the component scope.
+useEventListener(window, 'keydown', handleEscKey)
+useEventListener(
+  window,
+  GUIDED_ONBOARDING_RESUME_REQUESTED_EVENT,
+  handleGuidedOnboardingResumeRequested as EventListener
+)
+
+void ensureStartupWelcomeState()
+
+watch(
+  () =>
+    [isStartupRouteReady.value, route.name, draftStore.pendingStartDeeplink?.token ?? 0] as const,
+  () => {
+    void activatePendingStartDeeplink()
+  },
+  { immediate: true }
+)
+
+onMounted(() => {
+  // Ensure icons are loaded (load asynchronously, can happen in parallel with store init)
+  void ensureIconsLoaded()
+
+  // Start shell-critical data directly from the main window so it does not depend on settings.
+  void initAppStores()
+  setupMcpDeeplink()
+  setupAppIpcRuntime()
+
+  watch(
+    () => activeTab.value,
+    (newVal) => {
+      router.push({ name: newVal })
+    }
+  )
+
+  watch(
+    () => route.fullPath,
+    (newVal) => {
+      const pathWithoutQuery = newVal.split('?')[0]
+      const newTab =
+        pathWithoutQuery === '/'
+          ? (route.name as string)
+          : pathWithoutQuery.split('/').filter(Boolean)[0] || ''
+      if (newTab !== activeTab.value) {
+        activeTab.value = newTab
+      }
+      // Close artifacts page when route changes
+      artifactStore.hideArtifact()
+    }
+  )
+
+  // Listen for changes to current conversation
+  watch(
+    () => sessionStore.activeSessionId,
+    () => {
+      // Close artifacts page when switching conversations
+      artifactStore.hideArtifact()
+    }
+  )
+})
+
+// Clear timers and event listeners before component unmounts
+onBeforeUnmount(() => {
+  if (errorDisplayTimer) {
+    clearTimeout(errorDisplayTimer)
+    errorDisplayTimer = null
+  }
+
+  cleanupAppIpcRuntime()
+  cleanupMcpDeeplink()
+})
+</script>
+
+<template>
+  <div
+    data-testid="app-root"
+    class="flex flex-col h-screen"
+    :class="isWinMacOS ? 'bg-window-background' : 'bg-background'"
+  >
+    <AppBar />
+    <div class="flex flex-row h-0 grow relative overflow-hidden px-px py-px" :dir="langStore.dir">
+      <div class="flex flex-row w-full h-full">
+        <WindowSideBar></WindowSideBar>
+
+        <!-- Main content area -->
+        <div
+          data-testid="app-main"
+          class="flex h-full min-h-0 flex-1 min-w-0 flex-col overflow-hidden rounded-tl-xl border-l border-t border-black/20 bg-background dark:border-white/10"
+        >
+          <div class="min-h-0 flex-1">
+            <RouterView v-if="isStartupRouteReady" />
+          </div>
+        </div>
+      </div>
+    </div>
+    <!-- Global message dialog -->
+    <MessageDialog />
+    <McpSamplingDialog />
+    <!-- Global Toast notifications -->
+    <Toaster :theme="toasterTheme" />
+    <SelectedTextContextMenu />
+    <TranslatePopup />
+    <SpotlightOverlay />
+    <!-- Global model check dialog -->
+    <ModelCheckDialog
+      :open="modelCheckStore.isDialogOpen"
+      :provider-id="modelCheckStore.currentProviderId"
+      @update:open="
+        (open) => {
+          if (!open) modelCheckStore.closeDialog()
+        }
+      "
+    />
+  </div>
+</template>
