+++ vitest.config.renderer.ts
@@ -19,7 +19,6 @@ export default defineConfig({
     alias: {
       '@': resolve('src/renderer/src'),
       '@api': resolve('src/renderer/api'),
-      '@browser': resolve('src/renderer/browser'),
       '@shadcn': resolve('src/shadcn'),
       '@shared': resolve('src/shared'),
       vue: 'vue/dist/vue.esm-bundler.js'
