+++ vitest.config.ts
@@ -40,7 +40,6 @@ export default defineConfig({
             // Renderer process aliases (match electron.vite.config.ts renderer config)
             { find: '@/', replacement: resolve('src/renderer/src/') + '/' },
             { find: '@api', replacement: resolve('src/renderer/api') },
-            { find: '@browser', replacement: resolve('src/renderer/browser/') },
             { find: '@shared', replacement: resolve('src/shared') },
             { find: '@shadcn', replacement: resolve('src/shadcn') },
             { find: 'electron', replacement: resolve('test/mocks/electron.ts') },
